[backend] Updatemployee

+++ Backend/src/main/java/com/example/demo/repository/EmployeesRepo.java
@@ -3,7 +3,7 @@ package com.example.demo.repository;
 import com.example.demo.model.Employees;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface EmployeesRepo extends MongoRepository<Employees, Integer> {
+public interface EmployeesRepo extends MongoRepository<Employees, Long> {
     Employees findByUsername(String username);
     Employees findByEmpId(String empId);
 }
